[posix-app] log filename when die
This PR aims at adding the filename when die. Other enhancements:

* Enable PLATFORM region by default for posix-app
* Disable warning for zero variadic macro arguments
* Use VerifyOrExit to implement other assert utilities
* Do not call fprintf() and print exit code in assert utility

To print dying message to stderr, add `-v` when launching the app.
To get the exit code, use `echo $?` just after the app died.

+++ tests/unit/test_hdlc.cpp
@@ -517,7 +517,7 @@ void TestEncoderDecoder(void)
 
 uint32_t GetRandom(uint32_t max)
 {
-    return rand() % max;
+    return static_cast<uint32_t>(rand()) % max;
 }
 
 void TestFuzzEncoderDecoder(void)
